\encoding{UTF-8}
\name{summary.CTTB}
\alias{summary.CTTB}
\title{Summarize results for "CTTB" objects}
\description{Summarize results generated by the \code{CTTB} method.}

\usage{\method{summary}{CTTB}(result, ...)
}

\arguments{
\item{CTTB}{a generated \code{CTTB} object.}
}

\value{
  Data frames storing results from the \code{CTTB} function.
}

\author{
  Cyrus Samii (\email{cds2083@nyu.edu}), New York University.

  Ye Wang (\email{yewang@unc.edu}), UNC-Chapel Hill.

  Aaron Junlong Zhou (\email{jlzhou@nyu.edu}), Tecent America
}

\references{
  Athey, Susan, Julie Tibshirani, and Stefan Wager. (2019).
  Generalized Random Forests.
  \emph{The Annals of Statistics, 47}(2): 1148–78.

  Chernozhukov, Victor, Denis Chetverikov, Mert Demirer, Esther Duflo, Christian Hansen, Whitney Newey, and James Robins. (2018).
  Double/Debiased Machine Learning for Treatment and Structural Parameters.
  \emph{The Econometrics Journal, 21}(1): C1–68.

  Lee, David S. (2009).
  Training, Wages, and Sample Selection: Estimating Sharp Bounds on Treatment Effects.
  \emph{The Review of Economic Studies, 76}(3): 1071–1102.

  Samii, Cyrus, Ye Wang, and Junlong Aaron Zhou. (2025).
  Generalizing Trimming Bounds for Endogenously Missing Outcome Data Using Random Forests.
  \emph{Political Analysis, First View}: 1-15.
}

\examples{
library(CTTB)
data(simData)

result <- CTTB(data = dat, seed = NULL, Y = "Y", D = "D", S = "S", X = c(names(dat)[c(2:12)]), W = NULL, Pscore = "Ps", cv_fold = 5, aggBounds = 1, LeeBounds = 1, cBounds = 0, X_moderator = NULL, cond.mono = FALSE)

summary(result)
}


